// src/app/modules/academicSemester/academicSemester.interface.ts
import { Model } from 'mongoose'

export type IMonths =
  | 'January'
  | 'February'
  | 'March'
  | 'April'
  | 'May'
  | 'June'
  | 'July'
  | 'August'
  | 'September'
  | 'October'
  | 'November'
  | 'December'

export type IAcademicSemesterTitles = 'Autumn' | 'Summer' | 'Fall'

export type IAcademicSemesterCodes = '01' | '02' | '03'

export type IAcademicSemester = {
  title: IAcademicSemesterTitles
  year: number
  code: IAcademicSemesterCodes
  startMonth: IMonths
  endMonth: IMonths
}

export type AcademicSemesterModel = Model<
  IAcademicSemester,
  Record<string, unknown>
>